Pradnya/pyproject.toml
[project]
name = "mental-health-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-core>=1.38.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.4",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.0",
    "twilio>=9.10.0",
    "uvicorn>=0.40.0",
]
